+++ internal/datasource_replace.go
@@ -1,4 +1,4 @@
-package util
+package internal
 
 import (
 	"context"
