+++ internal/nixane/collect_test.go
@@ -1,4 +1,4 @@
-package internal
+package nixane
 
 import (
 	"testing"
