+++ internal/nixane/encode.go
@@ -1,4 +1,4 @@
-package internal
+package nixane
 
 import (
 	"fmt"
