+++ internal/nixane/fs.go
@@ -1,4 +1,4 @@
-package internal
+package nixane
 
 import (
 	"io/fs"
