+++ internal/provider_test.go
@@ -1,4 +1,4 @@
-package util
+package internal
 
 import (
 	"testing"
