+++ internal/resource_register_test.go
@@ -1,4 +1,4 @@
-package util
+package internal
 
 import (
 	"testing"
