+++ main.go
@@ -3,11 +3,11 @@ package main
 import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 
-	"github.com/poseidon/terraform-provider-util/util"
+	"github.com/poseidon/terraform-provider-util/internal"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
-		ProviderFunc: util.Provider,
+		ProviderFunc: internal.Provider,
 	})
 }
